Remove espressif macs

+++ esp32_marauder/WiFiScan.h
@@ -116,7 +116,7 @@ esp_err_t esp_wifi_80211_tx(wifi_interface_t ifx, const void *buffer, int len, b
 
 struct ssid {
   String essid;
-  int channel;
+  uint8_t channel;
   int bssid[6];
   bool selected;
 };
